Created script for Settings Menu - Not tested in build
I have created the script for the options menu to include a resolution dropdown, a fullscreen toggle and a volume slider. I have scripted the functionality for the dropdown and toggle and created the basis for the volume slider which is commented out with a debug statement in place currently.