Started on possible Patient Spawner Script and issues with animations
I started to work on a script to get the patients to spawn in the scene using tags on the spawn points. I still need to figure out how to populate the patients array and how to instantiate them. I also tried to work on animations to be used for background patients but I had issues getting the mixamo animations to work.