Fixes to script and testing of spawner script
I have added a start function to my UI script currently to ensure that the scene's timescale is correct and that is isn't affecting the animation for the models but this wasn't the issue. I also worked on testing the spawner script so now the spawn point array populates on play and I have manually added models into the patients array. Still haven't got the instantiate working.