Updated Pause Menu script + added setting menu script - not all working
I updated the pause menu script to add functionality for the terminology UI button. I also implemented the setting menu script into the game. Current the fullscreen toggle works and the volume slider works in editor as I haven't connected up a mixer due to the lack of audio in the game but will work on this after it has been added. I also found on build that the slider doesn't work. The functionality for the resolution dropdown is working BUT I will need to redo this as the text does not change due to how I set it up.