Fixed Settings Menu in Tut and Main Menu Scenes + added audio mixer
I fixed the resolution dropdown in the main menu and tutorial scenes and connected everything up. I also added an audio mixer and connected it up to the audio sources in the game scene and set the slider to full otherwise this caused issues with the audio playing in the scene. I did not work in the main scene due to last having seen Harlem working in it and seeing no pushes made recently. I have not tested the current game in a build yet.