Updated Start Menu functionality (animations need to be redone)
Added functionality to the  Main Menu scene which now leads the player to the free roam scene and player data scene. I don't know if I have linked to the correct scene for the player data so this can be changed if needed. Due to making new buttons the animation for the buttons needs to be redone.