Credit Scene retrieved
I went through the old text file from my edit on the scene as that is the second last version before the merge errors were found and I copied that version of the text file and pasted it into the text file for the current credits scene. I tested the scene and it now works.